+++ modules/grafana/default.nix
@@ -1,33 +1,61 @@
-{ config, ... }: {
+{ config, self, ... }:
+let inherit (self.common.networking.services) grafana loki prometheus;
+in {
   networking.firewall.allowedTCPPorts =
     [ config.services.grafana.settings.server.http_port ];
 
   services.grafana = {
     enable = true;
 
+    provision = {
+      enable = true;
+
+      datasources.settings.datasources = [
+        {
+          name = "Loki";
+          type = "loki";
+          access = "proxy";
+          url =
+            "${loki.protocol}://${loki.ipv4}:${builtins.toString loki.port}";
+        }
+        {
+          name = "Prometheus";
+          type = "prometheus";
+          access = "proxy";
+          url = "${prometheus.protocol}://${prometheus.ipv4}:${
+              builtins.toString prometheus.port
+            }";
+        }
+      ];
+    };
+
     settings = {
+      "auth.anonymous".enabled = false;
       alerting.enabled = false;
       analytics.reporting_enabled = false;
-      "auth.anonymous".enabled = false;
       panels.disable_sanitize_html = true;
+
       security = {
         # TODO: change before deploy
-        admin_user = "admin";
         admin_password = "test";
+        admin_user = "admin";
       };
-      unified_alerting.enabled = true;
+
       server = {
-        http_addr = "0.0.0.0";
         # domain = "grafana.rovacsek.com";
         # root_url = "https://${name}.${domain}/";
         enable_gzip = true;
         # TODO: change prior to deploy
         enforce_domain = false;
         # TODO: change prior to deploy
-        protocol = "http";
-        http_port = 3002;
+        http_addr = "0.0.0.0";
+
+        inherit (grafana) protocol;
+        http_port = grafana.port;
       };
+
       users.auto_assign_org_role = "Editor";
+      unified_alerting.enabled = true;
     };
   };
 }